Variety of changes
I changed surveys to hopefully diseapper temporarily when the user submits, also I added the expandible list to have recent notifications its not working fully as of yet as my version of the notification functions is behind. I also added some arrays to the user's interface to get some of this functionality working.